requirements.txt
fastapi>=0.68.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
uvicorn>=0.15.0
python-multipart>=0.0.5  